🔨 REFACTOR (exhibition): move tag prop declaration into view
