+++ src/pages/Exhibitions/Exhibitions.tsx
@@ -1,32 +1,12 @@
 // Libraries
 import React from "react";
 import { Link } from "react-router-dom";
-
-import { SectionProps } from "./Section";
-import { RoleProps } from "./Role";
+import { HTMLAttributes } from "react";
+import ExhibitionModel from "./Exhibition/models/ExhibitionModel"
 // Styles
 import "./style.css";
 
-enum VisibilityTypes {
-  Public,
-  Exclusive,
-  Private,
-}
-
-interface ExhibitionProps {
-  UUID: string;
-  host: string;
-  coHost: string[];
-  title: string;
-  description: string;
-  keywords: string;
-  sections: SectionProps[];
-  roles: RoleProps[];
-  openingDate: Date;
-  closingDate: Date;
-  visibility: VisibilityTypes;
-  invitedAttendees: string[];
-}
+type ExhibitionProps = ExhibitionModel & HTMLAttributes<HTMLElement>;
 
 const Exhibitions: React.FC = () => (
   <div className="exhibitions">
